Change to prevent gcc testsuite from terminating or1ksim
On running the gcc testsuite various tests would be compiled and
run in sequence with the same executable name. For an as-yet unknown
reason gdb does not reload the symbols correctly causing a
breakpoint placed on _exit to be placed in the position of _exit
in the previously tested executable. With varying optimisation
levels this caused the breakpoint to sometimes be missed allowing
execution to hit l.nop 0xc, aborting or1ksim silently.

For now this workaround simply makes gdb call 'file' with no
parameters to clear the symbols before the next file load.