Added scripts to sort failures in the gcc test log
sort_failures.py sorts according to regular expressions of common error messages. filter_log.sh uses this to create a blacklist of files, as well as a manifest of expected failures to be used with validate_failures.py in the gcc testsuite. The resultant blacklist and manifest will need manual tweaking, as they will include all failed tests in the log regardless of whether they are genuine failures.